chore: moved telemetry to argilla.utils chore: used get_telemetry_client method
