+++ src/argilla/server/apis/v0/handlers/users.py
@@ -21,13 +21,13 @@ from pydantic import parse_obj_as
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from argilla.server import models
-from argilla.server.commons import telemetry
 from argilla.server.contexts import accounts
 from argilla.server.database import get_async_db
 from argilla.server.errors import EntityAlreadyExistsError, EntityNotFoundError
 from argilla.server.policies import UserPolicy, authorize
 from argilla.server.security import auth
 from argilla.server.security.model import User, UserCreate
+from argilla.utils import telemetry
 
 router = APIRouter(tags=["users"])
