+++ tests/conftest.py
@@ -27,13 +27,13 @@ from argilla.client.datasets import read_datasets
 from argilla.client.models import Text2TextRecord, TextClassificationRecord
 from argilla.client.sdk.users import api as users_api
 from argilla.datasets.__init__ import configure_dataset
-from argilla.server.commons import telemetry
-from argilla.server.commons.telemetry import TelemetryClient
 from argilla.server.database import Base, get_async_db
 from argilla.server.models import User, UserRole, Workspace
 from argilla.server.search_engine import SearchEngine, get_search_engine
 from argilla.server.server import app, argilla_app
 from argilla.server.settings import settings
+from argilla.utils import telemetry
+from argilla.utils.telemetry import TelemetryClient
 from fastapi.testclient import TestClient
 from opensearchpy import OpenSearch
 from sqlalchemy import create_engine
